Handling anchored routes
When a route is marked as not anchored (i.e. when defined with `mount`),
the translated routes lose this anchor marker.
The consequence of this, in my case, was the Resque web ui route not
being matched any more.

This fix provides the anchor option true/false value of the original
route to the translated ones.